Java implementation for transaction log reader/writer
Summary:
- Java implementation for transaction log reader/writer
- Unit tests
- Tests with randomly generated large transaction log files using MemFileSystem
- Bug fixes

Test Plan: Included unit tests.

Reviewers: peter

Reviewed By: peter

CC: justin, peter, tberman, serdar

Differential Revision: http://phabricator.thefactory.com/D1478

+++ java/src/main/java/com/thefactory/datastore/Batch.java
@@ -70,10 +70,10 @@ public class Batch {
     private int readLength() throws IOException {
         Slice in = asSlice();
         int offsetBefore = in.getOffset();
-        int ret = Msgpack.readRawLength(in);
+        long ret = Msgpack.readRawLength(in);
         int nBytesRead = in.getOffset() - offsetBefore;
         buffer.skipBytes(nBytesRead);
-        return ret;
+        return (int)ret;
     }
 
     private Slice readSlice() throws IOException {
